Started working on the camera.
The Player tank is now shaking when it moves. The camera moves in Late Update, and it uses Lerp to move, and the movement is in Fixed Update, so there shouldn't be any reason the player is shaking when moving. Also, none of the other object shakes, so it's not the camera. I've asked the professor for help, and I'll continue when he responds.